extract processor and make async
